Notify timer cancellation by calling delegate passing cancel enum
Instead of using semaphore, this is way more elegant. Plus the semaphore lead to
deadlocks (the stopToken was running onStop callbacks, waiting in the semaphore
for the task to be executed by the waiting thread, which was busy adding a new
timer to the same stopToken and waiting on the onStop callbacks to complete).