Windows: Fix duplicate concurrency_{globalStopSourcePointer,getLocalThreadExecutor} symbols
The `export` visibility causes these 2 special symbols to be exported
from every binary the concurrency library is linked statically into.
Now if we have a DLL exporting it, and the executable too, and the
executable is linked implicitly against the DLL (via .lib import library),
the linker complains about duplicate symbols.

So on Windows, it should generally be exported from a single binary
in the whole process. AFAICT, the `dynamicLoad` utility checks the
executable's exports only, so relies on that exporting binary being
the executable. This can be achieved via extra linker flags for the
executable; AFAIK, there's unfortunately no way to un-export specific
symbols in the linker command-line (which would alternatively allow
preventing the export via linker flags for the DLLs).