Trying to get connectHeap to play nice
Normally the state object, because it is on the stack, is destroyed
somewhere _after_ the receiver is called. With `connectHeap` this is
not the case, since the state lives on the heap.

Instead we call destroy manually to simulate this.

Also note that connectHeap can probably be used to escape references to
scoped sender/receivers, so its a hodgepot.

Anyway, its not working...