Fix: IOUring timer management
The iouring module was wrongfully propagating to a non-timeout wait from
a timeout submitAndWait under certain circumstances, leading to timeouts
never happening.

`dirtyTimers` need to be toggled even when no timers were triggered.
This is because the timingwheel implementation doesn't return the exact
time the next timer is to be triggered, but maxes that at 255 * tick size.

Test Plan: `dub test`.